move caching tests to separate file
Existing tests for --name-cache and --path-cache are moved to a separate
file.
